Fix bug with hash map keys
The hash map key was just an integer and the supported hash map key
types (string, boolean, integer) were all just naively casting their
internal value to an int when generating the hash key for the object
type. This meant bools and ints could easily cause collisions, since
true is casted to 1.

For example, the following would occur:
```
>> let map = {true: 1, 1: 2}
{true: 1}
```

This diff fixes it by making the hash key a string that encodes both
the type and the value of the Object to prevent collisions across
types.